requirements.txt
# The whole scientic test is run in python virtual environment therefore
# you need to install python=3.6.7, link for ubuntu users: https://www.digitalocean.com/community/tutorials/how-to-install-python-3-and-set-up-a-programming-environment-on-ubuntu-18-04-quickstart
# you will also need to upgrade pip to newest version or else you will get problems with matplotlib
# There can occur problems with running xgboost on Windows, because xgboost is not quite compatible with Windows
# so it is better to set up enviornemnt on Linux og MacOS

# With this little text file all the necessary requirements to run the
# project can be installed quickly in terminal using following command
# $ pip install -r requirements.txt
# or if you will use conda as your environment you may install it manually with conda commands each one separately because of the dependencies each one has
#
numpy
pandas
py-xgboost-gpu
#xgboost
py-xgboost-gpu
scikit-learn
tensorflow-gpu
tensorflow
keras
matplotlib
nltk
seaborn
tabulate
gensim
paramiko